"10:05am. I got up 20m ago. Let me see if I can go through vol 9 and then I will start.
So far, my impression of Dendro is that it is really good. Especially the characterization - Hell General acting like a cartoon villain and then being shown to be a 10 year old kid throwing a tantrum after getting BTFOd is quite credible. The author is quite talented to be able to show this. Dendro is more than just the MC being thrown in some fantasy world and being given powers - the border between the world is there, and it is getting used.

10:20am. Ok, focus me. Waste time in the right way.

10:40am. https://www.reddit.com/r/MachineLearning/comments/eq3da0/d_what_are_the_current_significant_trends_in_ml/

Right now, I am reading Dendro, but decided to check out this thread on a whim.

"Look into sparse evolutionary training, it’s using genetic algorithms to configure networks for learning. Also symbolic regression is starting to gain popularity again as more modern uses have been published last year. Type those in using google scholar and you’ll see some cool stuff. The SET technique above was published in Nature I believe. If you have troubles let me know I can find the papers for you ❤️"

This stuff is new to me, I'll definitely keep it in mind. I never really looked into symbolic regression for real either though the talk by Lipshuts that I watched way before I started this quest really sold me on its potential.

After everything is done, I will really be innovating in the direction of using evolution together with NNs.

https://www.reddit.com/r/MachineLearning/comments/eqdad9/d_a_sober_look_at_bayesian_neural_networks/

I'll look at this at a later date, let me get back to my vacationing.

1:25pm. Done with vol 9. I've really been indulging myself for the past 1.5 weeks.

I think that starting next week I will try taking things a bit more seriously. The way things are going is taking too long, so I get to testing the new language soon. I'll roll up my sleeves and put in a few extra hours if possible and get back to my old pace.

1:30pm. I really can't keep working on the language forever.

I need to get back into ML.

So for the near term, at least finishing the first half should be my main priority. Even if I decide to skip typechecking for some reason, I still need to beef up Spiral's compilation speeds. So that is work that absolutely needs to be done.

1:35pm. Just one more time. I will evolve the language one final time and then try again.

The way I was approaching ML reminds me of how in the past I'd approached roguelike games like Diablo. At that time the degree of caution needed to win at roguelikes was completely foreign to me.

Games like Angband I've tried getting into many, many times and every time I would get creamed and die. Back when I was a kid I liked using cheats, and I like exploring immersive worlds and collecting stuff. The challenge was secondary.

Me becoming capable of beating Angband, DCSS, DoomRL is a fairly recent phenomenon.

I guess when I decided to become a trader it brought a change in my mindset and approach towards the games. It is not just that I had a fundamentally different approach to these games - I did. It is also that I had changed my affinities - I started to enjoy playing in the new style. And that made a world of difference in my survivability.

It is really a pity that I could never get anywhere with trading. If you are aiming for longer term, beating the market is not too hard, but you'll never get anywhere with 3k which was my situation. I've seen good daytraders do it as a challenge, but they already had the necessary skills developed from before.

But that is a thing of the past now. If my AI plan succeeds, I will never have to trade again anyway, at least as far as financial instruments are concerned.

1:45pm. For the past 5 years I've been slowly building my skills. I've ignored all the shortcuts and did all the exercises.

I think I know a lot in both ML and PL now.

Just like with roguelikes, it is not that even back then I was too dumb to really beat them. It is just that I've been approaching ML too much like regular programming and am getting irate when I get pushback from reality.

Sure, I might not know how NNs and CFR really work. I cannot reason it out.

In the future there might be some genius who can do these things that I cannot.

1:50pm. But at a certain point one needs to accept that there are classes of algorithms that are simple, but due to various reasons (such as working in continuous spaces) cannot be reasoned about.

It is like - suppose I did not understand basic addition of natural numbers. Then I'd get obsessed about mastering the 'fundamentals', crack open the books and get down to the bare circuits.

I'd find that addition can be done in terms of various logic circuits and train myself to reason about them in those terms.

And yet, while this effort might gain me that reasoning ability, I would be no closer to understanding at all.

It is like a being with no visual system trying to understand geometry.

...

No doubt there are higher levels of intelligence not allowed to humans that might allow shedding light on these basic algorithms, but I can honestly say that I gave it my best shot. Whatever understanding I wanted to unearth with my 2019 effort is certainly beyond my ability.

Hard thinking and great effort will not suffice here.

1:55pm. Instead I will accept that what I had was simply naivety and change my approach to something more befitting of a person living in this era.

I'll think of ML as science itself, and think of backprop as nothing more than greedy improvement. I'll assume that there is no secret sauce - rather what I need to do is increase the roboustness of it in various domains through the sheer weight of my programming skill.

Forget about money and ML both.

What I need to do is find a way to make an expression of science. Reach the new perspective and grasp it. Program in a altogether different manner.

Leave the others to waste their time trying to fiddle with the rules and architectures.

This is in fact what I said I would do, but I got completely obsessed with gaining 'understanding'. The end result is that the more I learned, the further I move away from it.

2pm. If I moved with the assumption that algorithms were understanding itself - this is something that everybody deep down already believes and acts upon anyway, then I would have not gotten so completely lost. I would not have started crying blood and fuming fire as soon as I realized what a scam modern math is.

...

I really am deeply regretful that I could not realize all of this soon. Before I even stared work on ML, I should have realized that I do not really understand even much simpler algorithms that work in rational spaces. Had I done so, I would not have wasted so much time trying to drill a hole through reality.

I want redemption for this sin.

And the best way to do this is to dedicate myself wholly to programming once more.

This time, I won't ask for money, understanding, recognition or anything else from the art.

2:05pm. When I was a kid, I realized something deep in my core. At that point I did not know about the Singularity or anything like that, but I could tell - the era I lived in was special.

The games that can be played on the computer are special.

The world beyond the screen is the world the God only knows.

2:10pm. I need to do it. Failures and victories do not matter.

What is important is that I express my understanding of what understanding really is.

Do I really want to give up here and be a monkey for the rest of my life? Or do I want to act as one of the predecessors to the Inspired?

Things such as life can be left to lesser creatures. I have the boundary to break."