"``` type Env<'a> = { exists : Set<string>; free_vars : Dictionary<string,'a>; local : Map<string,'a> } ```
Uah, I just remembered that I forgot the counters again.

```
type Env<'a> = { exists : Set<string>; free_vars : Dictionary<string,'a>; local : Map<string,'a>; local_index : int ref }
```

Let me go with this.

```
type Env<'a> = { exists : Set<string>; free_vars : Dictionary<string,'a>; local : Map<string,'a>; local_index : int }
```

No wait, actually it should not be a ref.

2:10pm.

```
    let value_add_local a (env : LocEnv) =
        {env.value with
            local_index=env.value.local_index+1
            local=Map.add a (V env.value.local_index) env.value.local
            }
```

This is how I should add local vars.

```
    let type_add_local a (env : LocEnv) =
        {env.type' with
            local_index=env.type'.local_index+1
            local=Map.add a (TV env.type'.local_index) env.type'.local
            }
```

Might as well get this out of the way ahead of time as well.

Ah, crap I forgot to add to `exists` as well.

```
    let value_add_local a (env : LocEnv) =
        {env.value with
            exists=Set.add a env.value.exists
            local_index=env.value.local_index+1
            local=Map.add a (V env.value.local_index) env.value.local
            }

    let type_add_local a (env : LocEnv) =
        {env.type' with
            exists=Set.add a env.type'.exists
            local_index=env.type'.local_index+1
            local=Map.add a (TV env.type'.local_index) env.type'.local
            }
```

This is why these helpers are used. It is too easy to forget what needs to be done otherwise.

2:15pm. Ah, crap. I never realized it before, but env.value with does not at all do the same things as in Spiral. This is lame of F#.

It really is atrocious when it comes to records isn't it?

```
    let value_add_local a (env : LocEnv) =
        {env with value =
                    {env.value with
                        exists=Set.add a env.value.exists
                        local_index=env.value.local_index+1
                        local=Map.add a (V env.value.local_index) env.value.local
                        }
            }

    let type_add_local a (env : LocEnv) =
        {env with type' =
                    {env.type' with
                        exists=Set.add a env.type'.exists
                        local_index=env.type'.local_index+1
                        local=Map.add a (TV env.type'.local_index) env.type'.local
                        }
            }
```

Ugh, just bear with it.

2:25pm.

```
        | RawInl (a,b) ->
            let env' = fresh_env() |> value_add_local a
            let e, stack = value_prepass env b
            Inl(e,expr_data_of env env' stack), em
```

Well, this should be fine now. I thought it would be more complicated, but this should work.

```
        | RawForall (a,b) ->
            let env' = fresh_env() |> type_add_local a
            let e, stack = value_prepass env b
            Forall(e,expr_data_of env env' stack), em
```

Here is the `Forall` as well. It is quite similar to the `Inl`.

Let me continue going.

2:30pm. Focus me.

```
        | RawKeywordCreate (a,b) ->
            KeywordCreate()
```

Now comes `KeywordCreate`. I think I just need to use the `Array.mapFold` here.

2:45pm.

```
        | RawKeywordCreate (a,b) ->
            let b,sz =
                Array.mapFold (fun sz x ->
                    let x,sz' = value_prepass env x
                    x,max' sz sz'
                    ) em b
            KeywordCreate(keywords.To a,b),sz
```

Forgot that I need the keyword to int dictionary as well. But I've added it.

```
| RawRecordWith (a,b) -> failwith ""
```

This thing is next.

Really, this stuff is not that difficult. Let me do the entire prepass today. I just need to stop moping around and just plug away at it.

```
        | RawRecordWith (a,b) ->
            let a,sz = Array.mapFold (fun sz x -> let x,sz' = value_prepass env a in x,max' sz sz') em b
```

Actually, before I do this one, let me do the let. Let is the thing that increments the stack counter. Also...

```
let env' = fresh_env() |> value_add_local a
```

I made a mistake here. This thing should be value 1.

```
    let value_sz_add i (t,v) = t,v+i
    let type_sz_add i (t,v) = t+i,v
```

Ok...

```
        | RawInl (a,b) ->
            let env' = fresh_env() |> value_add_local a
            let e, sz = value_prepass env b
            Inl(e,expr_data_of env env' (value_sz_add 1 sz)), em
        | RawForall (a,b) ->
            let env' = fresh_env() |> type_add_local a
            let e, sz = value_prepass env b
            Forall(e,expr_data_of env env' (type_sz_add 1 sz)), em
```

Yeah, it really is easy to make this mistake. Let me do this a bit more abstractly.

```
    let vlet env a b =
        let e,sz = b (value_add_local a env)
        e, value_sz_add 1 sz
    let tlet env a b =
        let e,sz = b (type_add_local a env)
        e, type_sz_add 1 sz
```

Let me do it through this.

3:10pm. Had to take a tiny break.

```
        | RawInl (a,b) ->
            let env' = fresh_env()
            let e, sz = vlet env' a (fun env -> value_prepass env b)
            Inl(e,expr_data_of env env' sz), em
        | RawForall (a,b) ->
            let env' = fresh_env()
            let e, sz = tlet env' a (fun env -> value_prepass env b)
            Forall(e,expr_data_of env env' sz), em
```

Focus me, focus.

I've done this. What else is there.

3:25pm. You know what, screw propagating the size information like I am doing now. It would be far easier to just keep track of the max value.

```
type Env<'a> = { exists : Set<string>; free_vars : Dictionary<string,'a>; local : Map<string,'a>; local_index : int; local_index_max : int ref }
```

I am really running a huge risk making a mistake somewhere in the prepass when adding things to the environment.

3:30pm.

```
    let value_add_local a (env : LocEnv) =
        {env with value =
                    {env.value with
                        exists=Set.add a env.value.exists
                        local_index=env.value.local_index+1
                        local_index_max=max (env.value.local_index+1) !env.value.local_index_max
                        local=Map.add a (V env.value.local_index) env.value.local
                        }
            }

    let type_add_local a (env : LocEnv) =
        {env with type' =
                    {env.type' with
                        exists=Set.add a env.type'.exists
                        local_index=env.type'.local_index+1
                        local_index_max=max (env.type'.local_index+1) !env.type'.local_index_max
                        local=Map.add a (TV env.type'.local_index) env.type'.local
                        }
            }
```

Yeah, this is the right choice. It is easy to forget, but imperative programming is more powerful that the functional one in terms of style.

Previously I had to take great care to do those maxes, but here it is all done for me on its own.

Let me commit here, and I will redesign the things that I've done so far."