"6:55pm. Ok, it is possible to fix this with a few changes:
1) I will disallow local shadowing of globals.
2) I will remove existentials from the prepass.
3) Rather than memoizing the entire thing in RawInline, I will do something more along the lines of...

```
        | RawInline e -> // TODO: Completely broken.
            let e,env' =
                memoize dict_rawinline (fun _ ->
                    let env' = fresh_env'()
                    value_prepass env e, env'
                    ) e
            Inline(e,expr_data_of env env')
```

This will be sound.

7:10pm. I am thinking.

In typecase and type bindings, I think I am going to have to use specialized syntax to indicate that I want to do unification.

Imagine this.

```
match x with
| ((a : q) | a) ->
    typecase e with
    | q -> ...
```

In one situation the e will be bound to q, but in another it will fall through.

I was wrong. The right answer is not to propagate variables that exist forward.

```
match x with
| ((a : q) | a) ->
    typecase e with
    | bind q. q -> ...
```

Rather, I need something like this to indicate that I want to use unification or not.

7:20pm. I just haven't thought the situation around variant type bindings through enough it seems.

7:25pm. I am actually tempted to put in overtime at this juncture, but 7:25pm is really no time to be starting new things. If I feel that strongly about it, I just should go to bed early so I can start work early tomorrow.

Actually I am glad it turned out like this because the way semantics worked seemed awkward even when I started this.

It is really a pity I cannot shadow globals locally, but I will be able to deal with it."