"9am. I am up. I actually woke up earlier, but have been indulging myself in my imagination.
If there is a direction I want to change, it would be to take my imagination more seriously.

Deep down, I've been too laid back and skeptical.

This is reflected in the last 5 years. Spiral since its inception was always an expression of both my belief and skepticism in ML. But I did not channel this skepticism properly. A really common pattern in my explorations is that I would figure out some big optimization like KFAC, Zap, that cumulative eligibility trace update, and then find that they are local minima in terms of design. Even minor architectural changes break them.

In the end, I can only count of plain old backprop.

And starting from 2015, ever since the discrete optimization course, I did not really understand why anything other that point optimizations were necessary. Even now I do not have an absolute view I could point to, but nonetheless the hints are everywhere that diversity is the key to uncertainty estimation.

Back in 2015, I thought that there would be something smarter than multiple random inits, but I guess not.

9:10am. This time I really want to do so much more.

And conversely, this is the last time I will ever work on a language. I swear it. V0.2 is the last time I will ever put so much effort into this because honestly, I had enough. I do not want to end up in a situation where in 2030 I both still a pure human and still fiddling with this donkey shit.

Programs are crystallizations of understanding, so I am going to finish it here.

9:35pm. Ok, enough slacking. Let me see whether I can get some work done. I need to start work on partial evaluation. Just two more years of this and then I am done for good.

When v0.2 is done, I will have no more refactoring or compilation speed issues anymore. It needs to be done.

9:50am. Yeah, it is going to be hard going to start. Right now rather than programming I am posting on /sci/. For the past few months my frequency of posting has been way up. From once a year on average to a few times per week.

...Ah, let me chill for a while longer and then I will start."