"10:55am. I only got up 5m ago. Yeah, I'll definitely have that breakfast right away.
Today and tomorrow I won't be focusing too much on getting programming done as I have other things on my plate, but I'll try to get a little done.

https://blog.floydhub.com/gans-story-so-far/
Generative Adversarial Networks - The Story So Far

Yesterday I've been reading about GANs, and have spent a part of the night thinking about it. My usual habit would be to focus on the various issue with backprop and try to figure out how to fix them...but maybe that is the wrong way to go about it here.

For the first time I can feel it - maybe it is okay if the agent just dies during training. Maybe that is an intrinsic property of life.

My job isn't to fiddle with credit assignment. It is to break time and make the networks be capable of save/loading.

Deep down, I really am in awe at backprop. From a certain angle it really does all make sense and I cannot think of a better way to do credit assignment. It is both elegant, flexible and extremely resilient to attempts at displacing it.

11:05am. It absolutely is a brilliant idea to make the models continuous and then to calculate the effect of small changes. I'll give it that.

I am not going to worry or beat myself over not being able to figure out how to improve on that brilliant idea. Nature did it, but evolution's main trait its ability to emulate lateral thinking. Humans are much more focused in comparison, which has its benefits and weaknesses.

In the worst case, if the 20s turn out to be a bust and people can't do it, putting microscopes into the brain will clear things up eventually.

11:10am. I have to accept one thing though, every learning rule will have its weaknesses.

And what I will aim from here on out is to work around them at the meta level. My attempts to fix the bugs directly have all failed, so I will go for more inspired designs and hopefully make gains that way.

12:05pm. Done with breakfast.

Yeah, it is great that I can finally channel this feeling. I'd be really effective as ML user if I could stop myself from going down dead ends. I really needed this bitter medicine - sometimes the hardest lesson to learn is that something is irreducible. And sometimes reducing something even if possible results in a loss of structure, clarity and understanding.

12:25pm. Let me do the chores and then comes bath. After that I'll have the rest of the day to myself.

A bit by bit I am kindling this feeling into a flame. I will complete Spiral v0.2 and then I will be absolutely free to program the way I want.

I am going to ditch my nasty habits from the last two years and aim for my dream with what is my original power.

Without a doubt, all this work on PL has leveled me up in some ways, so even if I try to innovate, I can be sure that the things I will be doing will be out of reach for any kind of ML researcher. I am going to exploit this to it truest advantage.

I am going to stop messing around. When the time comes to make the ML library again, the first thing I will do replace the arrays in the allocator with weak arrays. I made a huge strategic blunder by not letting the GC manage memory. If necessary I will change compilation targets from .NET to something else if necessary in order to have that be robust.

12:35pm. This time there will be no more type leakiness, nor compilation slowdowns as the program gets big.

I will stop looking down on languages like C and take something good from them like the top level. I won't be arrogant and think that I can do without top-down typing either.

I am going to learn from the drumming that I've received and do things right this time.

Life is long and I still have time to do it properly.

The fundamentals of the situation have not changed. I am the one who has."