"11:20am. Today I will be taking a break as I went to sleep at 6am and haven't gotten a wink of rest. I am also hung over.
I am not going to be doing any good programming in this state, so I won't do it at all.

11:25am. The night was interesting. The experience really pushed me in a big way towards where I had already started going this month.

For the first time, I've looked back at my failures of the past 5 years from a confident perspective and it made me realize something.

Besides my main goal, I have really many contradictory desires which I've been dealing with by averting my eyes and suppressing them. One is protecting those close to me. As one might expect I do not exactly want to shove my own parents into the oven. I do not want to be heartless. I want to be honest with my desires.

But much like the human civilization was not exactly the best thing to happen to the animal kingdom, the next evolutionary step will be a pretty bad thing for >99.99% of all humans. The self improvement talent should not be a widespread thing.

11:35am. So causing this sort of situation to occur goes against that desire. I am really perplexed by this. It feels that I fell into the same trap as the rationalists except from a different angle.

11:40am. Maybe it is the drugs and the wine talking, but I've had a certain vision. I've neglected to consider something very obvious in predictions - unlike in say logic, contradictory desires are not necessarily errors in the program even though they feel that way subjectively. I do not know what exactly is going to happen post-Singularity, but it might turn out to be the case that contradictions themselves are the main source of inspiration. It is not necessarily the case that they need to be resolved.

The brain is hiding a lot of stuff that we do not know about in plain view. The way Singularity will play out pretty much all depends on algorithmic realities of intelligence.

11:45am. And one more thing - my vision is spinning here, so let me write it down quickly.

For the past 5 years I've been operating on the assumption that the brain is very stable. But if I think about it a bit more closely, there is no reason at all to make that assumption. There is no reason to think that it is not savescumming much like I intend to do from here on out.

If I 'died' a few times over the night, there is actually no way I could ever verify it. And the brain certainly has every incentive not to freak me out with the glitches in its programming, so that sort of phenomenon would never get noticed.

So much like with contradictory desires, I am going to make a simplifying assumption here as well since assumptions about what the brain is doing have an out-sized effect on ML research direction regardless of their truthfulness.

It might be wrong, but it is better strategically. If it turns out to be true and I assume the opposite, then I will have made a huge blunder by casting my net in the wrong algorithm design space. If it turns out to be false, and I assume it is true, then correcting that error will be really cheap.

So I haven't been playing the right way here up to now.

Lastly, I do not need to hide for all the deadends I've gone down and the lack of progress. Humans place great stock on momentum, but I feel that Inspired with a legacy of failure are more realistic than with one of success. Failing upward and going forth with the demeanor of a king regardless of the results is what a rational agent would do. It treat the pain of regret as a inspiration tap."