"https://eschew.wordpress.com/2020/01/26/undergrad-compilers-from-the-hive-mind/
Well, let me read this first. It is actually pretty interesting.

11:40am. "We would also like to underline the importance of models interpretation. By having the possibility of building a model visualization or extracting the set of rules that govern the SNN model, we are able to interpret the internal mechanisms behind the model learning. In contrast to other ANNs where it is hard to look into the network and figure out exactly what or how it has learned, some SNNs allow us to know about their learning process (Soltic & Kasabov, 2010)."

This is new to me. Why would SNNs be more interpretable?

12:20pm. Done with breakfast.

https://www.youtube.com/results?search_query=Spiking+Neural+Network&page=&utm_source=opensearch

A lot of recent stuff on SNNs is here. I really do not feel like programming right now. Let me instead do some studying.

Let me just say, that my problem is not necessarily ANNs communicating via floats as it is differentiability. Differentiability is the problem with everything.

Spiking NNs being spiking is secondary to me. The main draw here is that they are non-differentiable. Since I cannot figure out how to get rid of differentiability directly, it might be better to transport myself to a regime where I do not have access to it in the first place.

https://www.youtube.com/watch?v=WStnSDZ3ygA
Spiking Neural Networks: A Primer with Dr. Terrence Sejnowski - #317

Let me start with this thing."