"2:45pm. https://youtu.be/jhQgElvtb1s?t=240
What he is saying here is interesting. I do not really understand why SNNs would be more general than ANNs.

> Mike: There is a lot of confusion for those who do not know much about the neuromorphic space. An assumption that what we are trying to do is build better deep learning accelerator chips or something like that, and that's really not the primary point of our research. It is an understandable point of confusion because the spiking NN model is in many ways a generalization of what an ANN does therefore you can generally map standard NNs into a spiking algorithm, but that is not necessarily the best place how to use spiking NNs. And there is all kinds of other properties, it is not all about the spikes by any means, there are many different architectural principles that are related here.

3pm. https://youtu.be/jhQgElvtb1s?t=719

He says that if I am interested I should mail him. Well, this is probably meant for actual academics, but after Spiral is good enough and there aren't any toys out there that I can get my hands on I'll consider giving it a try.

https://youtu.be/jhQgElvtb1s?t=760
The Challenge: SNN algorithm discovery.

Yeah, this is what I want. I find backpropagation just so oppressive. I really want to know what the alternatives to it would be.

https://youtu.be/jhQgElvtb1s?t=888
New ideas guided by neuroscience

This stuff is new to me.

https://youtu.be/jhQgElvtb1s?t=1304
Loihi can be used to do path planning

https://youtu.be/jhQgElvtb1s?t=1399
Loihi can be used to do constraint solving.

Everything is batch size 1 which is what I need for RL.

3:50pm. "If you are into this, let me also recommend: [Mike Davies: Realizing the Promise of Spiking Neuromorphic Hardware](https://www.youtube.com/watch?v=jhQgElvtb1s). This is a presentation on Loihi, the neuromorphic chip at development at Intel.

At the start of the video he mentions there being some [controversy](https://www.zdnet.com/article/intels-neuro-guru-slams-deep-learning-its-not-actually-learning/) regarding neuromorphic chips, and here he gives some justification for their development. I think he does a very good job hyping it.

He makes the claim that neuromorphic chips being DL accelerators is a misconception and that spiking NNs generalize ANNs. He then elaborates that through several examples as implemented on the Loihi. It can be used to do things like path planning and solving constraint satisfaction problems which are important for biological brains.

The main advantage of Loihi besides its energy efficiency is that unlike GPUs and CPUs it works great with batch size of 1 as it is asynchronous and latency focused. Personally, when I tried online RL on the GPU I found them to be a massive disappointment as without batching the benefits of parallelism could not be manifested.

He states that using backprop with SNNs is low hanging fruit, and the main challenge going forward is moving beyond that towards more neuroscience inspired approaches."

Let me post this as a comment.

3:50pm. https://www.youtube.com/watch?v=_Xcdm6alfCY
Spiking Neural Networks Deep Dive

https://www.youtube.com/watch?v=4lY-oAY0aQU
MIT 6.S191: Biologically Inspired Neural Networks (IBM)

Let me stop here for a bit so I can do the chores.

4:25pm. I am back.

I've been thinking. I've made up my mind - Mike Davies is essentially right about everything. Loihi (or something like it) is what I should be angling Spiral for.

4:30pm. 2022 is only two years away. I am seriously going to aim that Spiral v0.2 will have a neuromorphic backend rather than a CPU one.

In fact, since after I am done with v0.2 I won't really be able to upgrade the rig, I am going to have to look for a job and hope that Spiral looks good enough as the sole item on my resume. Rather than beat around the bush, I should just email the guy directly and ask him if they are interested in hiring a PL guy...well, remotely anyway.

4:35pm. I don't think backprop is not learning, but I certainly hate it at this point so I am definitely on the same page as Mike now.

4:40pm. Every long effort needs a certain hope. In 2018 that was that I could find some way of making backprop work well enough to create a stable RL agent.

Loihi might as well be my future. Algorithms are necessary of course, but hardware is the backbone of ML.

Despite the math talent in the world, literally nobody in the 90s could have foreseen deep learning. Proof is in having a program and running it.

4:50pm. Neuromorphic hardware was a far off future in 2015, but is near at hand now.

Forget all of that nonsense about reversibility and using evolutionary algos, and aim for this one thing. Literally nobody is smart enough to design ML algorithms from the ground up. The Inspired will be able to do it, but I am really far below their level of capability.

I am not alone. Intelligence is in my brain. Intelligence is everywhere in nature. It just needs to be minded and refined into something usable. Neuroscience has been useless so far, but it will get its time to shine. They will get their chance to earn their paycheck.

4:55pm. I can make Spiral, but I can't imagine myself doing any significant algorithmic breakthroughs in the ML space.

...Ah, shit. I really will have to get a job. Well whatever. I'll thinking about that when it comes to that.

5pm. Let me finish today off with one more video.

https://www.youtube.com/watch?v=4lY-oAY0aQU
MIT 6.S191: Biologically Inspired Neural Networks (IBM)

Let me go with this.

...No that is not on neuromorphic chips. Ok, let me go for the other one. It is nearly 2h though.

5:20pm. Went through the slides. It is just really basic and long. Skip.

https://www.youtube.com/watch?v=gX9NqDuwTnA
Neuromorphic computing with emerging memory devices

Let me give this a chance.

I do not care if SNNs do not work as ANNs. I just want to run away from ANNs. The desire to create new algorithms and my inability to do so is what is giving me nightmares. It would really be far more productive to focus my efforts where I can make the most impact.

7:05pm. Done with lunch. Let me close for the day here. Tomorrow I will get on those join points.

I guess when I finish the language I can try Deep CFR on the CPU as an exercise, but otherwise my aim will be to break into the neuro programming field. Forget evolution or anything like that. I need firm principles in order to do proper engineering. Without principles I might as well be guessing.

I am going to track down some of the reference that Davies mentioned on mathematical formalization of some of the things they are doing.

I need to gather my resolve. I need a firm vision in order to have a belief in the future.

There is just one way to do this, it is to have patience. There isn't a genius who can figure out learning. The community as a whole will be one to make the hardware and to infer the processes of the brain's functioning."