"1:10pm. Done with breakfast. Next are chores and chilling.
2:05pm. Done with that.

2:10pm. Now it is time to start, but let me focus inwardly for a bit.

I am thinking how I would talk the guy into hiring me 1-2 years from now, but that is the wrong mindset to be in.

It does not really matter what happens to Spiral. What matters is that I use my full power to reach my goals.

2:15pm. I am so rattled. I really do not feel like programming.

I said that thing about wanting to try evolagos, and now I am aiming for neuromorphic chips.

Yeah, there is definitely pain of not being able to aim towards my goals directly.

But it just goes to show how weak the trust in my ideas were that once I saw something even remotely realistic, I immediately ditch them.

2:20pm. Instead of working today, why don't I turn off and step away from the screen for a while?

Yeah, that is what I will do.

I just came up with the neurochip goal yesterday and presented it today here.

But as a result, now I am changing my direction again. Yesterday I said I would do Deep CFR, and now I suddenly do not feel like it because it would be sloppy work.

2:25pm. I really, really wanted to figure it out, but I knew that it would be unrealistic from the start. Even though I knew, it does not stop it from hurting. I just had to show my determination. I just had to test myself at how far I can go. I just had to take the responsibility for doing it myself.

Patience - a viable strategy, but a painful word for me.

I feel like I have a responsibility to make the Singularity come as soon as possible. I took it upon myself.

But now that I know the task of understanding learning is beyond the capacity of humans, the smart thing to do would be to change strategies and just like I am letting the work of creating hardware to others, to let the neuroscientists figure out the secrets of spiking NNs.

2:30pm. I am absolutely sure. I've tried my best during 2018-2019 and I know for certain that I cannot beat nature's design. I am supposedly a good programmer, but I am nothing much when the big picture is considered. Those lofty heights of programming skill are surely reserved for the post-human entities of the future who I want to be one of.

2:35pm. It is ironic that human technology compared to nature is more primal. It is one that taps into the core virtues of reality. Maybe that is because humans are weak, so they have to do it properly in order to ensure that it works.

Two years...

Two years...

Two years...

What will happen in two years? These chips will be out...and there should be something to make good use of them. Yesterday in one of the videos the speaker said that bee brains were realistic and that human brains are too complex, but he is not being realistic.

If bee brains are here, then human brains won't be too far behind.

2:40pm. The truth is at this point the Singularity is making me sick. Here I am trying to squeeze some kind of a plan here into a timeframe where the long term does not give me a risk of death - nor do I want my family members to just slip through the cracks of time, but the real world is going at its own pace.

2:45pm. Just where are neuromorphic chips now?

Where will they be in 2030 for example?

Pretty much all ML advancements are just buzzing around the hardware. Hardware is what acts as a catalyst to everything else.

2:55pm. https://www.extremetech.com/computing/295043-intels-neuromorphic-loihi-processor-scales-to-8m-neurons-64-cores

8m neurons compared to the brains 100b. Also Loihi has 1k synapses for every neuron so I'll assume this applies here as well.

"“With the Loihi chip we’ve been able to demonstrate 109 times lower power consumption running a real-time deep learning benchmark compared to a GPU, and 5 times lower power consumption compared to specialized IoT inference hardware,” said Chris Eliasmith, co-CEO of Applied Brain Research and professor at University of Waterloo. “Even better, as we scale the network up by 50 times, Loihi maintains real-time performance results and uses only 30 percent more power, whereas the IoT hardware uses 500 percent more power and is no longer real-time.”"

109 times lower power consumption than a GPU.

Upgrade the process a little, figure out the memristors, and 100b neurons on a chip sounds doable. Of course the brain will have a lot more complex architecture, is analog, so it might be much more powerful pound for pound.

But still...

Let me just update the review. A bit. I need to highlight just how humiliated I am at not being able to take advantage of the great programmability advantages on present hardware.

3:10pm. "I feel shame very strongly over this state of affairs.

I've lost, but it is not to other programmers or mathematicians. Rather if I have lost to somebody, it would be nature. I have such strong beliefs and conviction and yet despite the huge advantages that being able to program the hardware directly provides I am not able to convert that potential into actual power. But that does not change that the potential is still there waiting to be tapped."

Yeah, I am mad at the mathematicians as well, but being mad at nature is where the benefits will be.

"If I am to be mad or to hate the lack of progress, the place I should direct that effort is definitely nature. It is there that this sort of dedication will have merit."

3:20pm. I definitely felt the hate when I wrote this.

3:25pm. Let me go back to what I was writing.

I need to focus on this vision. In 2030 and most likely even sooner, the bee brain tier of power and skill should definitely be here. This is the worst case scenario, and even in it I should have enough leeway to move properly with the insights derived from brain research initiatives.

3:30pm. This was actually my mindset back before 2014 when I was having fun with futurism.

I need to get back to it. Forget about causing the Singularity by some fixed date. The self improvement loop will swallow me when the time comes.

I do not owe it to anybody to do it in 2022 or whatever.

If it happens by 2030 then that is fine. If it happens by 2040 then that is fine too. I will do my utmost - not to hold out that long, but to have fun in the interim.

That is what matters the most.

3:30pm. I need to regain the spark towards the future. I need to have joy in making Spiral.

I need to feel like I am making real progress, not just so I can make some agent or whatever, but for its own sake.

3:35pm. Now as I said, I would let me step away from the screen and I will just take a nap for a few hours.

I want to dwell on this feeling. When it bursts, I will come back and resume work on the language.

I vow one thing no matter what. I will never again clown around by saying I want some agent, or get frustrated when results inevitably do not arrive at the schedule I would like. I am absolutely through with that.

There is only one reason to make Spiral - because I am the only one who can.

And because I am the best."